\name{gpssmR-package}
\alias{gpssmR-package}
\alias{gpssmR}
\docType{package}
\title{\packageTitle{gpssmR}}
\description{\packageDescription{gpssmR}}
\section{Package Content}{\packageIndices{gpssmR}}
\author{\packageAuthor{gpssmR}}
\section{Maintainer}{\packageMaintainer{gpssmR}}
\keyword{package}
